% Generated by roxygen2 (4.0.1): do not edit by hand
\name{adjFactor}
\alias{adjFactor}
\title{Adjust Factors}
\usage{
adjFactor(jtids = NULL, vars = NULL, region, adjFactorFile = NULL)
}
\arguments{
\item{jtids}{stock symbols}

\item{vars}{High/low/open/close price for a stock}

\item{region}{region for stock symbols}

\item{adjFactorFile}{adjFactor File}
}
\value{
df
}
\description{
Adjust Factors
}
\details{
Adjust Factors
}
\author{
Weilin Lin
}

