% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{any_unnamed}
\alias{any_unnamed}
\title{Check for any unnamed parameters in call.}
\usage{
any_unnamed(call)
}
\arguments{
\item{call}{A call (list) object.}
}
\value{
\code{TRUE} if atleast one parameter is not named.
}
\description{
Check for any unnamed parameters in call.
}
