% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.data.frame.DataFrame}
\alias{as.data.frame.DataFrame}
\title{Convert \code{DataFrame} to \code{data.frame}}
\usage{
\method{as.data.frame}{DataFrame}(df)
}
\arguments{
\item{df}{A \code{DataFrame} object.}
}
\value{
A \code{data.frame} object.
}
\description{
Convert \code{DataFrame} to \code{data.frame}
}
