% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.data.table.DataFrame}
\alias{as.data.table.DataFrame}
\title{Convert \code{DataFrame} to \code{data.table}}
\usage{
\method{as.data.table}{DataFrame}(x)
}
\arguments{
\item{x}{A \code{DataFrame} object.}
}
\value{
A \code{data.table} object.
}
\description{
Convert \code{DataFrame} to \code{data.table}
}
