% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_n.R
\name{as_n}
\alias{as_n}
\title{Wrapper around as.numeric()}
\usage{
as_n(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric \code{vector}.
}
\description{
Wrapper around as.numeric()
}
\examples{
 \dontrun{

 as_n(rpois(10, 5))
 
 }


}
