% Generated by roxygen2 (4.0.1): do not edit by hand
\name{backtest.bondFuture}
\alias{backtest.bondFuture}
\title{backtest main function for the bondfuture}
\usage{
backtest.bondFuture(cfgLst)
}
\arguments{
\item{cfgLst}{configuration file}
}
\value{
list
}
\description{
backtest main function for the bondfuture
}
\details{
backtest main function for the bondfuture
}
\author{
Weilin Lin
}

