% Generated by roxygen2 (4.0.1): do not edit by hand
\name{backtest.demo.bondFuture}
\alias{backtest.demo.bondFuture}
\title{Demo for bondfuture HFT}
\usage{
backtest.demo.bondFuture()
}
\value{
data frame
}
\description{
Demo for bondfuture HFT
}
\details{
Demo for bondfuture HFT
}
\author{
Weilin Lin
}

