% Generated by roxygen2 (4.0.1): do not edit by hand
\name{backtest.loadData.bondFuture}
\alias{backtest.loadData.bondFuture}
\title{load data for bond future HFT strategy}
\usage{
backtest.loadData.bondFuture(cfgLst)
}
\arguments{
\item{cfgLst}{configuration list object for the backtesting}
}
\value{
list
}
\description{
load data for bond future HFT strategy
}
\details{
load data for bond future HFT strategy
}
\author{
Weilin Lin
}

