% Generated by roxygen2 (4.0.1): do not edit by hand
\name{backtest.loadData}
\alias{backtest.loadData}
\title{backtest the stock selection strategy}
\usage{
backtest.loadData(startDate = NULL, endDate = NULL, cfgLst)
}
\arguments{
\item{startDate}{startDate of backtest}

\item{endDate}{endDate of backtest}

\item{cfgLst}{configuration list object for backtest}
}
\value{
cfgLst contains market data
}
\description{
backtest the stock selection strategy
}
\details{
backtest the stock selection strategy
}
\author{
Weilin Lin
}

