% Generated by roxygen2 (4.0.1): do not edit by hand
\name{backtest.optimizedRebalance}
\alias{backtest.optimizedRebalance}
\title{more complex optimizer for backtesting}
\usage{
backtest.optimizedRebalance(hldDf, alphaDf, dataDf, univVec, cfgLst,
  longIDs = NULL, reduceIDs = NULL)
}
\arguments{
\item{hldDf}{holding data frame}

\item{alphaDf}{alpha data frame}

\item{dataDf}{data frama}

\item{univVec}{universe vector}

\item{cfgLst}{configuration list}

\item{longIDs}{longIDs}

\item{reduceIDs}{reduceIDs}
}
\value{
data frame
}
\description{
more complex optimizer for backtesting
}
\details{
more complex optimizer for backtesting
}
\author{
Weilin Lin
}

