% Generated by roxygen2 (4.0.1): do not edit by hand
\name{backtest}
\alias{backtest}
\title{backtest main function}
\usage{
backtest(cfgLst)
}
\arguments{
\item{cfgLst}{configuration list}
}
\value{
cfgLst
}
\description{
backtest main function
}
\details{
backtest main function
}
\author{
Weilin Lin
}

