% Generated by roxygen2 (4.0.1): do not edit by hand
\name{backtest.simpleRebalance}
\alias{backtest.simpleRebalance}
\title{simple rebalance function for backtesting}
\usage{
backtest.simpleRebalance(hldDf, alphaDf, dataDf, univVec, cfgLst)
}
\arguments{
\item{hldDf}{holding data frame}

\item{alphaDf}{alpha data frame}

\item{dataDf}{data frame}

\item{univVec}{univ vector}

\item{cfgLst}{configuration}
}
\value{
cfgLst
}
\description{
simple rebalance function for backtesting
}
\details{
simple rebalance function for backtesting
}
\author{
Weilin Lin
}

