% Generated by roxygen2 (4.0.1): do not edit by hand
\name{backtest.summary.bondFuture}
\alias{backtest.summary.bondFuture}
\title{backtest summary for the bondfuture}
\usage{
backtest.summary.bondFuture(cfgLst)
}
\arguments{
\item{cfgLst}{configuration list object}
}
\value{
data frame and plot
}
\description{
backtest summary for the bondfuture
}
\details{
backtest summary for the bondfuture
}
\author{
Weilin Lin
}

