% Generated by roxygen2 (4.0.1): do not edit by hand
\name{backtest.summary}
\alias{backtest.summary}
\title{backtest summary for the backtest}
\usage{
backtest.summary(cfgLst, startDate = NULL, endDate = NULL,
  pnlVar = "betaAdjPnL", extraTC = 0.002, plotToFile = TRUE,
  navDf = NULL)
}
\arguments{
\item{cfgLst}{configuration list}

\item{startDate}{startDate of summary}

\item{endDate}{endDate of summary}

\item{pnlVar}{pnlVar}

\item{extraTC}{extra trading cost}

\item{plotToFile}{determine whether to plot to pdf for summary}

\item{navDf}{daily summary data frame}
}
\value{
list with summary by different frequency
}
\description{
backtest summary for the backtest
}
\details{
backtest summary for the backtest
}
\author{
Weilin Lin
}

