% Generated by roxygen2 (4.0.1): do not edit by hand
\name{beta}
\alias{beta}
\title{reads beta}
\usage{
beta(startDate, endDate = startDate, jtids = NULL, vars = NULL,
  region = "CN", inDir = NULL, UPDN = 0, source = "MySQL", ...)
}
\arguments{
\item{startDate}{start Date}

\item{endDate}{end Date}

\item{jtids}{stock symbols}

\item{vars}{varibles}

\item{region}{region}

\item{...}{other parameters}
}
\value{
beta
}
\description{
reads beta
}
\details{
reads beta
}
\author{
Weilin Lin
}

