% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bloomberg.cumRetPlot}
\alias{bloomberg.cumRetPlot}
\title{cunmRet plot for asset price}
\usage{
bloomberg.cumRetPlot(indexNames, ChiNames = NULL, startDate = NULL,
  endDate = NULL, monthLabel = TRUE, ...)
}
\arguments{
\item{indexNames}{indexNames}

\item{ChiNames}{ChiNames}

\item{startDate}{startDate}

\item{endDate}{endDate}

\item{monthLabel}{whether use month label}

\item{...}{others}
}
\value{
gg plt
}
\description{
cunmRet plot for asset price
}
\details{
cunmRet plot for asset price
}
\author{
Weilin Lin
}

