% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bloomberg.gg.daily}
\alias{bloomberg.gg.daily}
\title{using ggplot to find season pattern}
\usage{
bloomberg.gg.daily(name, startDate = NULL, endDate = NULL,
  market = "market", ...)
}
\arguments{
\item{startDate}{startDate of the data}

\item{endDate}{endDate of the data}
}
\value{
xts object
}
\description{
using ggplot to find season pattern
}
\details{
using ggplot to find season pattern
}
\author{
Yishuo Deng
}

