% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bloomberg.gg.month}
\alias{bloomberg.gg.month}
\title{using ggplot to find season pattern}
\usage{
bloomberg.gg.month(name, startDate = NULL, endDate = NULL, df = NULL)
}
\arguments{
\item{startDate}{startDate of the data}

\item{endDate}{endDate of the data}
}
\value{
xts object
}
\description{
using ggplot to find season pattern
}
\details{
using ggplot to find season pattern
}
\author{
Weilin Lin
}

