% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bloomberg.macro.key}
\alias{bloomberg.macro.key}
\title{key stats for macro}
\usage{
bloomberg.macro.key(name, startDate = NULL, endDate = NULL, ...)
}
\arguments{
\item{name}{}

\item{startDate}{}

\item{endDate}{}

\item{...}{}
}
\value{
data frame
}
\description{
key stats for macro
}
\details{
key stats for macro
}
\author{
Weilin Lin
}

