% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bloomberg.macro}
\alias{bloomberg.macro}
\title{Macro data from bloomberg}
\usage{
bloomberg.macro(name, startDate = NULL, endDate = NULL, verbose = TRUE,
  ...)
}
\arguments{
\item{name}{ticker of bloomberg}

\item{startDate}{startDate of data}

\item{endDate}{endDate of data}

\item{verbose}{whether to plot and print}

\item{...}{}
}
\value{
data frame and plot
}
\description{
Macro data from bloomberg
}
\details{
Macro data from bloomberg
}
\author{
Weilin Lin
}

