% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bloomberg.market.getData}
\alias{bloomberg.market.getData}
\title{bloombereg get data}
\usage{
bloomberg.market.getData(names, startDate = NULL, endDate = NULL, ...)
}
\arguments{
\item{names}{names of indicators}

\item{startDate}{}

\item{endDate}{}

\item{...}{}
}
\value{
xts object
}
\description{
bloombereg get data
}
\details{
bloombereg get data
}
\author{
Weilin Lin
}

