% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bloomberg.market.key}
\alias{bloomberg.market.key}
\title{market key data}
\usage{
bloomberg.market.key(startDate = NULL, endDate = NULL, ...)
}
\arguments{
\item{startDate}{}

\item{endDate}{}

\item{...}{}
}
\value{
data frame
}
\description{
show key stats of market
}
\details{
show key stats of market
}
\author{
Weilin Lin
}

