% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bloomberg.market}
\alias{bloomberg.market}
\title{get market data}
\usage{
bloomberg.market(name, startDate = NULL, endDate = NULL, addSMA = FALSE,
  ifPlot = TRUE, ...)
}
\arguments{
\item{name}{symbol of the bloomberg ticker}

\item{startDate}{startDate of data}

\item{endDate}{endDate of data}

\item{addSMA}{if add simple moving average line to the plot}

\item{ifPlot}{if plot to the window}

\item{...}{}
}
\value{
data frame and plot
}
\description{
get market information data
}
\details{
get market information data
}
\author{
Weilin Lin
}

