% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bloomberg.plot.monthly}
\alias{bloomberg.plot.monthly}
\title{plot macro by month}
\usage{
bloomberg.plot.monthly(df, ...)
}
\arguments{
\item{df}{}

\item{...}{}
}
\value{
data frame
}
\description{
plot macro by month
}
\details{
plot macro by month
}
\author{
Weilin Lin
}

