% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bloomberg.pmi}
\alias{bloomberg.pmi}
\title{PMI panel}
\usage{
bloomberg.pmi(startDate = NULL, endDate = NULL, ...)
}
\arguments{
\item{startDate}{}

\item{endDate}{}

\item{...}{}
}
\value{
xts object
}
\description{
PMI panel
}
\details{
PMI panel
}
\author{
among group
}

