% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bloomberg.update.CB}
\alias{bloomberg.update.CB}
\title{update convertible bond market index data}
\usage{
bloomberg.update.CB()
}
\description{
update convertible bond market index data
}
\details{
update convertible bond market index data
}
\author{
Weilin Lin
}

