% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bloomberg.update.derivative}
\alias{bloomberg.update.derivative}
\title{update derivative indicators}
\usage{
bloomberg.update.derivative()
}
\description{
update derivative indicators
}
\details{
update derivative indicators
}
\author{
Weilin Lin
}

