% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bloomberg.update.wind}
\alias{bloomberg.update.wind}
\title{update macro and market data}
\usage{
bloomberg.update.wind()
}
\description{
update macro and market data
}
\details{
update macro and market data
}
\author{
Weilin Lin
}

