% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bondMain}
\alias{bondMain}
\title{collect the information fo variety of bonds}
\usage{
bondMain(jtids = NULL, vars = NULL, region = "CN", inFile = NULL)
}
\arguments{
\item{jtids}{symbols for bonds}

\item{vars}{variables in data}

\item{region}{region for bonds}

\item{inFile}{data source}
}
\value{
df
}
\description{
This is the scripts to collect the information of variety of bonds
}
\details{
This is the scripts to collect the information of variety of bonds
}
\author{
Weilin Lin
}

