% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_beta_param.R
\name{calc_beta_param}
\alias{calc_beta_param}
\title{Calculates Beta Distribution Parameters from its Mean and Variance Parameters}
\usage{
calc_beta_param(mu, var)
}
\arguments{
\item{mu}{mean}

\item{var}{variance}
}
\value{
named vector with alpha/beta parameters.
}
\description{
Calculates Beta Distribution Parameters from its Mean and Variance Parameters
}

