% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calCumRet}
\alias{calCumRet}
\title{caculates the cumulative return}
\usage{
calCumRet(jtids, startDates, endDates, region = "CN")
}
\arguments{
\item{jtids}{symbols}

\item{startDates}{start dates}

\item{endDates}{end dates}

\item{region}{region}
}
\value{
cumRet
}
\description{
caculates the cumulative return
}
\details{
cumRet is computed use the as (endDate closePrice)/ (startDate closePrice) -1 with proper CACS adjustments
}
\author{
Weilin Lin
}

