% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cashFlow}
\alias{cashFlow}
\title{This is the function to calculate cash flow in deliyable set}
\usage{
cashFlow(jtid = NULL, vars = NULL)
}
\arguments{
\item{jtid}{symbols of stocks or bonds}

\item{vars}{variables in data}
}
\value{
a dataframe includs the cash flow of jtid
}
\description{
This is the function to calculate cash flow in deliyable set
}
\details{
This is the function to calculate cash flow in deliyable set
}
\author{
Weilin Lin
}

