% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandling.r
\name{cFactor}
\alias{cFactor}
\title{Paste factores}
\usage{
cFactor(x, y)
}
\arguments{
\item{x}{Factor x}

\item{y}{Factor y}
}
\value{
Factor
}
\description{
Paste factores
}
\details{
Paste factores
}
\author{
Jochen Kruppa
}
