% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{chisq.stars}
\alias{chisq.stars}
\title{Graphical interpretation of the chi-square test p-value}
\usage{
chisq.stars(
  x,
  y = NULL,
  correct = TRUE,
  p = rep(1/length(x), length(x)),
  rescale.p = FALSE,
  simulate.p.value = FALSE,
  B = 2000
)
}
\arguments{
\item{x}{a numeric vector or matrix. x and y can also both be factors.}

\item{y}{a numeric vector; ignored if x is a matrix. If x is a factor, y should be a factor of the same length}

\item{correct}{a logical indicating whether to apply continuity correction when computing the test statistic for 2 by 2 tables: one half is subtracted from all |O - E| differences; however, the correction will not be bigger than the differences themselves. No correction is done if simulate.p.value = TRUE.}

\item{p}{a vector of probabilities of the same length of x. An error is given if any entry of p is negative.}

\item{rescale.p}{a logical scalar; if TRUE then p is rescaled (if necessary) to sum to 1. If rescale.p is FALSE, and p does not sum to 1, an error is given.}

\item{simulate.p.value}{a logical indicating whether to compute p-values by Monte Carlo simulation.}

\item{B}{an integer specifying the number of replicates used in the Monte Carlo test.}
}
\value{
A character vector.
}
\description{
Returns a graphical interpretation of the chi-square test p-value
}
\details{
The function returns a graphical interpretation of a chi-square test p-value, under the form of a variable number of asterisks : (ns) means p >= 0.05; * means p < 0.05; ** means p < 0.01 ; *** means p < 0.001.
}
\examples{
## From Agresti(2007) p.39
M <- as.table(rbind(c(762, 327, 468), c(484, 239, 477)))
dimnames(M) <- list(gender = c("F", "M"), party = c("Democrat","Independent", "Republican"))
chisq.test(M))  # Prints test summary with chisq.test() function
chisq.stars(M) # Prints graphical test summary with stars
}
\seealso{
\code{\link[stats]{chisq.test}}
}
