% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandling.r
\name{chunks}
\alias{chunks}
\title{Split data into chunks}
\usage{
chunks(x, size)
}
\arguments{
\item{x}{Vector}

\item{size}{Size of chunks}
}
\value{
List
}
\description{
Split data into chunks
}
\details{
Split data into chunks
}
\examples{
d <- rpois(73, 5)
chunks(d, size = 10)
}
\author{
Jochen Kruppa
}

