% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanNAs.R
\name{cleanNAs}
\alias{cleanNAs}
\title{cleanNAs Function}
\usage{
cleanNAs(df)
}
\arguments{
\item{df}{A DataFrame}
}
\description{
Given a dataframe, this changes all NA, NULL, and empty string values to -9s.
}
\examples{
test <- data.frame(x = c(NA, NA, 1))
cleanNAs(test)
}
\keyword{dataframe}
