% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Codebook}
\usage{
codebook(data)
}
\arguments{
\item{data}{input data frame.}
}
\value{
A PDF file stored in the working directory.
}
\description{
This function builds PDF codebooks from data frames.
}
\examples{
data(mtcars)
codebook(mtcars)
}
