% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWideToLong.R
\name{convertWideToLong}
\alias{convertWideToLong}
\title{Convert Wide To Long}
\usage{
convertWideToLong(data, length)
}
\arguments{
\item{data}{this is the data.frame to be transformed}

\item{length}{the number of repeated blocks}
}
\description{
This function converts patient data in wide form to long form
}
\examples{
convertWideToLong()
}
\keyword{dataTools}
