% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cramer.v.stars}
\alias{cramer.v.stars}
\title{Graphical interpretation of Cramer's V}
\usage{
cramer.v.stars(tab)
}
\arguments{
\item{tab}{table on which to compute  Cramer's V test.}
}
\description{
Returns a graphical interpretation of the Cramer's V.
}
\details{
The function returns a graphical interpretation of the Cramer's V measure of the strength of assocation between two nominal variables, under the form of a variable number of asterisks : no star means V < 0.05; * means V >= 0.05; ** means V >= 0.15 ; *** means V < 0.30.
}
\examples{
## From Agresti(2007) p.39
M <- as.table(rbind(c(762, 327, 468), c(484, 239, 477)))
dimnames(M) <- list(gender = c("F", "M"), party = c("Democrat","Independent", "Republican"))
cramer.v(M))  # Computes Cramer's V with cramer.v() function from questionr package
cramer.v.stars(M) # Prints graphical Cramer's V summary with stars
}
\seealso{
\code{\link[questionr]{cramer.v}}
}
