% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CTDBond}
\alias{CTDBond}
\title{read CTDBond from our database}
\usage{
CTDBond(contract = NULL, region = "CN")
}
\arguments{
\item{contract}{contract}

\item{region}{region}
}
\value{
a data frame
}
\description{
read CTDBond from our database
}
\details{
read CTDBond from our database
}
\author{
Weilin Lin
}

