% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cumRet}
\alias{cumRet}
\title{Read cumRet from our Database}
\usage{
cumRet(startDate, endDate = startDate, jtids = NULL, vars = NULL, region,
  inDir = NULL, ...)
}
\arguments{
\item{startDate}{The startDate of data}

\item{endDate}{The endDate of data}

\item{jtids}{Stock or coupon symbols}

\item{vars}{Variables in data}

\item{region}{region of stock or coupon}

\item{inDir}{File direction to read data}

\item{...}{Other parameters needed}
}
\value{
dailyDF
}
\description{
Read cumRet from our Database
}
\details{
Read cumRet from our Database
}
\author{
Weilin Lin
}

