% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumsum1}
\alias{cumsum1}
\title{Cumulative sum of uninterrupted '1'}
\usage{
cumsum1(x)
}
\arguments{
\item{x}{integer vector to be used}
}
\value{
integer vector
}
\description{
Cumulative sum of uninterrupted '1'
}
\details{
Returns a vector whose elements are the 
cumulative sums of uninterrupted entries equal to 1 in an integer vector
}
\examples{
 \dontrun{

 x <- c(0,1,1,1,0,0,1,1,1,0,0)

 cumsum1(x)

 } 



}
