% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools-package.R
\docType{package}
\name{datatools-package}
\alias{datatools-package}
\alias{datatools}
\title{Simple tools I tend to use a lot when working with data in R}
\author{
Moritz Marbach \email{moritz.marbach@gess.ethz.ch}
}
