% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools-package.R
\docType{package}
\name{datatools}
\alias{datatools}
\title{datatools: A package of functions for historical and social science quantitative research}
\description{
The datatools package provides a set of functions for historical and social science quantitative research
}
