% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{Compute descriptive statistics}
\usage{
descriptives(data, ...)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{...}{variable names (unquote) for which to compute the descriptive statistics}
}
\value{
\code{data.frame} with summary statistics.
}
\description{
Compute descriptive statistics
}
\details{
If no a single variable name is passed to the function, it computes 
descriptive statistics across all variables. Before computing all 
variables are coerced via \code{as.numeric()}. Missing values are ignored 
(\code{na.rm=TRUE}).
}
\examples{
 \dontrun{

 x <- rpois(100,10)
y <- rnorm(100,0, 2)
z <- rexp(100, 1)
dat <- data.frame(x=x, y=y, z=z)
 descriptives(dat, z, x)
 descriptives(dat)

 }

}
