% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dir0}
\alias{dir0}
\title{Lazy alias for dir() function}
\usage{
dir0(...)
}
\arguments{
\item{...}{}
}
\value{
file list
}
\description{
Lazy alias for dir() function
}
\details{
Lazy alias for dir() function
}
\author{
Jochen Kruppa
}
