% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_na.R
\name{encode_na}
\alias{encode_na}
\title{Encode missing values as category and/or add missing data indicator}
\usage{
encode_na(df)
}
\arguments{
\item{df}{\code{data.frame}}
}
\value{
\code{data.frame}
}
\description{
A function useful to analyze missing data patterns.
}
\details{
Variables of class logical, character, factor, and ordered factor: Convert missings to a separate category ("(Missing)"). Character variables are converted to a factor variable. 
All other variables: Impute 0 and add a missing data indicator to data frame ("[varname]_na").
}
\examples{
df <- data.frame(a=1:5, b=LETTERS[1:5])
df[['a']][1] <- NA
df[['b']][4] <- NA
encode_na(df)


}
