% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_string.R
\name{encode_string}
\alias{encode_string}
\title{Encode a string vector into a set of numeric vectors}
\usage{
encode_string(string, method = "osa", k = 2, ...)
}
\arguments{
\item{string}{a string vector.}

\item{method}{the name of the distance metric.}

\item{k}{the maximum dimension of the space which the string is to be represented in.}
}
\value{
\code{data.frame}
}
\description{
Computes the distance matrix for a string and then applies multidimensional scaling to this distance matrix.
}
\details{
The method for computing distance, can be any of the string distance metrics implemented as part of the \code{stringdist} package (see \code{\link[stringdist]{stringdist-metrics}} for a list).
}
\examples{
a <- c("Berlin", "Hamburg", "München", "Köln") 
encode_string(a, method='cosine', k=3)


}
\seealso{
\code{\link[stringdist]{stringdist-metrics}}
}
