% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_score.R
\name{err_score}
\alias{err_score}
\title{Computes various scores useful for evaluating prediction accuracy}
\usage{
err_score(actual, predicted, slc = TRUE, m = "rmse")
}
\arguments{
\item{actual}{numeric vector with known values}

\item{predicted}{numeric vector with predicted values}

\item{slc}{logical vector indicating which values to use (default: all of them)}

\item{m}{score, can be either "rmse" (root mean squared error), 
"mad" (mean absolute error) or "cor" (Pearon's correlation).}
}
\value{
scalar score
}
\description{
Computes various scores useful for evaluating prediction accuracy
}
\details{
Missing values are ignored, ie. the score is computed based on 
complete cases.
}
\examples{
 \dontrun{


 a <- rnorm(10)
 b <- rnorm(10)
 
 err_score(a,b)

 } 



}
