% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_typing_err.R
\name{est_typing_err}
\alias{est_typing_err}
\title{Estimates the Expected Number of Errors in a Data Entry Job}
\usage{
est_typing_err(S, F, prior = "postmen", quantity = "mean")
}
\arguments{
\item{S}{size of the evaluation sample}

\item{F}{number of errors in the evaluation sample}

\item{prior}{can be either 'postmen', 'bad_postmen' or a named vector with prior values. See below for details.}

\item{quantity}{can be either 'mean', '95q' or 'param'. See below for details.}
}
\value{
value or vector of the requested quanty/ies.
}
\description{
Estimates the Expected Number of Errors in a Data Entry Job
}
\details{
The function calcualtes (Bayesian posterior) estimates for the expected number of errors (\code{quantity='mean'}) or the 
upper bound of the 95\% (Bayesian) credible interval (\code{quantity='95q'}) in the population of entered values when a 
random sample of \code{S} entered values contains \code{F} errors. For a discussion on the type of inference, see Winkler 
et al. (2002).

The calculations are done assuming a Beta-distribution prior over the proportion of errors for all entered values. The 
users must supply hyper-parameters for the Beta-prior. The function provides two informed default hyper-parameter sets. 
Both induce an informed prior and are formed from data about postmen's performance in typing post codes as reported in 
Baddeley/Longman (1978). 

The first prior (\code{prior='postmen'}) says that the a-priori the proportion of errors relative to the total number of entered 
values is about 1.4\% with a variance 0.7. The second prior (\code{prior='bad_postmen'}) is more conservative saying that 
it is about 2.1\% with a variance of 1.1. Users can also supply their own prior using a named vector that contains the alpha 
and beta hyper-parameters for the Beta prior (the output of \code{\link{calc_beta_param}} can be used).
}
\examples{
 \dontrun{

 est_typing_err(S=20, F=0, prior='postmen', quantity='95q')

 Example from Winkler et al. (2002) p.3:

 est_typing_err(167, 0, prior=c('alpha'=0.042, 'beta'=27.96), quantity='param' ) 

 } 



}
\references{
Baddeley, A. D., and D. J. A. Longman. 1978. "The influence of length and frequency of training session on the rate of learning to type." Ergonomics 21(8), 627-635.

 Winkler, Robert L. and Smith, James E. and Fryback, Dennis G. 2002. "The Role of Informative Priors in Zero-Numerator Problems: Being Conservative Versus Being Candid". The American Statistician 56(1), 1-4.
}
\seealso{
\code{\link{calc_beta_param}}
}
