% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandMultipleChoice.R
\name{expandMultipleChoice}
\alias{expandMultipleChoice}
\title{expandMultipleChoice}
\usage{
expandMultipleChoice(data, x, values, pattern = "", prefix = x,
  labels = values, sep = ".")
}
\arguments{
\item{data}{A data.frame containing the vector data that requires expansion}

\item{x}{Name of variable in 'data' containing vector needing expansion}

\item{values}{Vector of string values used to create concatenated string
response}

\item{pattern}{Pattern used to separate values in the concatenated string.
Default is "" for concatenated strings with no separator.}

\item{prefix}{Prefix to names of newly created variables}

\item{labels}{Vector of names to use for columns of resulting data.frame.
If not specified, columns are named using 'values'}

\item{sep}{to separate 'prefix' from 'labels' in the names of newly
created variables}
}
\value{
A data.frame containing the newly created variables expanded from
    input data
}
\description{
Function to expand response to a more than one answer multiple choice
question coded as a concatenated string (fix to CSPro multiple choice
questions)
}
\examples{

#

}
