% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extend_body}
\alias{extend_body}
\title{Extend the body of a function.}
\usage{
extend_body(fun, with, position = 1L)
}
\arguments{
\item{fun}{The function of which the body is to be extended.}

\item{with}{A language object to be inserted into the function body of \code{fun}.}

\item{position}{The position where \code{with} is to be inserted. Either 1 (before) or 2 (after) the existing body.}
}
\value{
A new function with the extended body.
}
\description{
Extend the body of a function.
}
\details{
A function must have a body in order to be able to extend it.
Primitive functions do not have a function body and trying to extend their body will result in an error.
See chapter 6.2.2 Primitive functions in https://adv-r.hadley.nz/functions.html.
}
\examples{
mean_verbose <- extend_body(mean, quote(message("Calculating mean!")), 1L)
mean_verbose(1:5)
}
