% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPatients.R
\name{extractPatients}
\alias{extractPatients}
\title{Extract Patients}
\usage{
extractPatients(ds, px)
}
\arguments{
\item{ds}{Parent data.frame to extract patient subset from.}

\item{px}{Vector of patient identifiers to extract from the dataset.}
}
\value{
A subset of the data.frame ds will be returned with only the patients
  listed in the px vector. The first column will be renamed "SID" as a result
  of this function.
}
\description{
Data.frames are subsetted using a vector of patients.
}
\details{
This function extracts a subset of patient data from a larger dataset. It
assumes the first column of the data.frame \code{ds} contains alphanumeric
subject identifications (\code{SID}) to be filtered; this column will be
renamed "SID" as a result of this function if it is not already.
}
\examples{
dataset <- data.frame(c(1:20), LETTERS[1:20], letters[1:20])
names(dataset) <- c("SID", "cov1", "cov2")
output <- extractPatients(dataset, 4:8)
output

}
\author{
Samuel Callisto \email{calli055@umn.edu}
}
