% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastx.R
\name{fastq_quality_control_old}
\alias{fastq_quality_control_old}
\title{Main fastq quality function}
\usage{
fastq_quality_control_old(inFile, tmpDir, p = 20, q = 50, leading = 10,
  trailing = 10, minlength = 50, illumninaclip = "TruSeq3")
}
\arguments{
\item{inFile}{Fastq infile}

\item{tmpDir}{Where to store the temporal files. Files will be
removed automatically}

\item{p}{Minimum percent of bases that must have [-q] quality.}

\item{q}{Minimum quality score to keep (phred)}

\item{leading}{Numeric. Remove leading low quality or N bases
(below quality 10 [default])}

\item{trailing}{Numeric. Remove trailing low quality or N bases
(below quality 10 [default])}

\item{minlength}{Min read length to keep. 50 [default]}

\item{illumninaclip}{Remove the adapter from illumina by TruSeq2
or TruSeq3 protocol}

\item{outFile}{Fastq outfile}
}
\description{
Main fastq quality function [deprecated]
}
\details{
Main fastq quality function. Calls the function
fastq_quality_filter and fastq_trimmer. Stores files in a tempDir.
}
\author{
Jochen Kruppa
}

