% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastx.R
\name{fastq_quality_filter}
\alias{fastq_quality_filter}
\title{Function to quality filter fastq files}
\usage{
fastq_quality_filter(inFile, outFile, q = 20, p = 50)
}
\arguments{
\item{inFile}{Fastq infiles}

\item{outFile}{Qc checked output file}

\item{q}{Minimum quality score to keep (phred)}

\item{p}{Minimum percent of bases that must have [-q] quality.}
}
\description{
Function to quality filter fastq files
}
\details{
Function to quality filter fastq files
}
\author{
Jochen Kruppa
}
