% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastx.R
\name{fastq_trimmer}
\alias{fastq_trimmer}
\title{Trimmomatic wrapper}
\usage{
fastq_trimmer(inFile, outFile, leading = 10, trailing = 10,
  minlength = 50, illuminaclip, log_file)
}
\arguments{
\item{inFile}{Fastq infile}

\item{outFile}{Fastq outfile}

\item{leading}{Numeric. Remove leading low quality or N bases
(below quality 10 [default])}

\item{trailing}{Numeric. Remove trailing low quality or N bases
(below quality 10 [default])}

\item{minlength}{Min read length to keep. 50 [default]}

\item{illuminaclip}{Remove the adapter from illumina by TruSeq2
or TruSeq3 protocol}
}
\description{
Trimmomatic wrapper
}
\details{
Trimmomatic wrapper. Will automatically idenitfy paied or single
reads, if they are named 'R1' and 'R2'. Not needed in single end
case.
}
\author{
Jochen Kruppa
}

