% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileDate.R
\name{fileDate}
\alias{fileDate}
\title{Append Date to Filename}
\usage{
fileDate(prefix = "_", suffix = ".csv")
}
\arguments{
\item{prefix:}{string to append before the date; defaults to "_"}

\item{suffix:}{string to be appended after the date, defaults to ".csv"}
}
\value{
character string containing prefix, date, and suffix.
}
\description{
fileDate is a small helper function which produces a
string containing today's date. The output is
intended to be appended to files which are frequently
iterated for organizational purposes.
}
\examples{
write.csv(iris, fileDate("iris_"))

}
\author{
Samuel Callisto \email{calli055@umn.edu}
}
