% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterC.R
\name{filterC}
\alias{filterC}
\title{Filter IGNORE=C}
\usage{
filterC(ds)
}
\arguments{
\item{ds}{longform dataset with IGNORE=C in first column}
}
\description{
This function filters out commented rows from a NONMEM-style
dataset, and removes the comment column for plotting.
}
\examples{
filterC()
}
\keyword{dataTools}
