% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstnamer-package.R
\docType{package}
\name{firstnamer}
\alias{firstnamer}
\alias{firstnamer-package}
\title{firstname: A package of functions using first name databases}
\description{
The firstnamer package provides a set of functions using first name databases
}
