% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstRow.R
\name{firstRow}
\alias{firstRow}
\title{Extract First Row}
\usage{
firstRow(ds, C = FALSE)
}
\arguments{
\item{ds}{longform dataset with patient ID in first column}

\item{C}{indicates inclusion of C column; if true, ignores first column}
}
\description{
This function extracts the first row for each patient from
a longform dataset, assuming patient identifiers are located
in the first column of the dataset.
}
\examples{
firstRow()
}
\keyword{dataTools}
