% Generated by roxygen2 (4.0.1): do not edit by hand
\name{futureMain}
\alias{futureMain}
\title{get future contract information and write to local csv files}
\usage{
futureMain()
}
\value{
data
}
\description{
get future contract information and write to local csv files
}
\details{
get future contract information and write to local csv files
}
\author{
Weilin Lin
}

