% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender.R
\name{gender}
\alias{gender}
\title{Predict genders from first names}
\usage{
gender(firstname, year_min = 1900, year_max = 2017, freq = FALSE)
}
\arguments{
\item{firstname}{first name as a character string.}

\item{year_min}{starting year of the period over which the prediction is computed.}

\item{year_max}{ending year of the period over which the prediction is computed.}

\item{freq}{return the probability of the first name being male}
}
\value{
The predicted genders based on the proportions of males and females with the input first names. Possible values are either "male" or "female". NAs are returned when the input first names are unknown in the database.
If freq is set to TRUE, the function returns the probabilities of the first names being male.
}
\description{
This function predicts the genders of a vector of first names.
}
\examples{
gender(c("Baptiste", "Henriette")
}
\seealso{
\code{\link{gender_unique}}, \code{\link{is_male}}, \code{\link{is_female}}
}
