% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_bt_depth}
\alias{get_bt_depth}
\title{Read bitcoin depth data}
\usage{
get_bt_depth(starttime = "2013-12-08 10:22:32",
  endtime = "2013-12-09 00:00:00", vars = "*")
}
\arguments{
\item{starttime}{The starttime of Data}

\item{endtime}{The endtime of Data}

\item{vars}{variables in data}
}
\value{
An dataframe
}
\description{
Read bitcoin depth data
}
\details{
Read bitcoin depth data
}
\author{
Yishuo Deng
}

