% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_bt_trade}
\alias{get_bt_trade}
\title{Read bitcoin trade data}
\usage{
get_bt_trade(starttime = "2013-12-09 00:00:00",
  endtime = "2013-12-09 00:25:30", vars = "*")
}
\arguments{
\item{starttime}{The starttime of Data}

\item{endtime}{The endtime of Data}

\item{vars}{variables in data}
}
\value{
An dataframe
}
\description{
Read bitcoin trade data
}
\details{
Read bitcoin trade data
}
\author{
Yishuo Deng
}

