% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCTD}
\alias{getCTD}
\title{Get CTD information}
\usage{
getCTD(startDate, endDate, market = "IB", contract, onlyCTD = TRUE,
  ifplot = TRUE)
}
\arguments{
\item{startDate}{}

\item{endDate}{}

\item{market}{}

\item{contract}{}

\item{onlyCTD}{}

\item{ifplot}{}
}
\value{
data frame and plot
}
\description{
Get CTD information
}
\details{
Get CTD information
}
\author{
Weilin Lin
}

