% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiHo.r
\name{getHoursMinutes}
\alias{getHoursMinutes}
\title{Wrapper for hours minutes}
\usage{
getHoursMinutes(time)
}
\arguments{
\item{time}{}
}
\value{
Hours:Minutes
}
\description{
Wrapper for hours minutes
}
\details{
Wrapper for hours minutes
}
\author{
Jochen Kruppa
}

