% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIO.r
\name{getObject}
\alias{getObject}
\title{ReadRDS wrapper}
\usage{
getObject(object)
}
\arguments{
\item{object}{a desired object name}
}
\value{
the object
}
\description{
Get a object back
}
\details{
Get a object back, which was saved as a RDS
}
\author{
Jochen Kruppa
}
