% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandling.r, R/utility.R
\name{gli}
\alias{gli}
\alias{gli}
\title{Small wrapper for glimpse()}
\usage{
gli(...)

gli(...)
}
\arguments{
\item{...}{}

\item{...}{}
}
\value{
glimpse()
}
\description{
Small wrapper for glimpse()

Alias for glimpse
}
\details{
Small wrapper for glimpse()

Alias for glimpse
}
\author{
Jochen Kruppa

Jochen Kruppa
}
