% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gower_dist.R
\name{gower_dist}
\alias{gower_dist}
\title{Computes The Extended Gower Distance Of Two Data Sets}
\usage{
gower_dist(
  data,
  query,
  return_index = FALSE,
  n_min = 1L,
  return_min = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{\code{data.frame}}

\item{query}{\code{data.frame}}

\item{return_index}{logical if TRUE return the index of the minimum distance}

\item{n_min}{integer number of values with smallest distance to be returned}

\item{return_min}{logical if the computed distances for the indices should be returned}

\item{verbose}{return information about variables}
}
\value{
\code{data.frame}
}
\description{
Wrapper around \code{\link[VIM]{gowerD}} with a simplified API.
}
\details{
\code{\link[VIM]{gowerD}} treats all variables as numeric if no type is supplied by the user.  \code{gower_dist} guess the variable type from the class name of the variable.
}
\examples{
data <- tibble(a=1:5, b=LETTERS[1:5])
data[['a']][1] <- NA
data[['b']][4] <- NA

query <- tibble(a=1:5, b=LETTERS[6:10])
gower_dist(data,query)


}
\seealso{
\code{\link[VIM]{gowerD}}
}
