% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs.R
\name{gs_household}
\alias{gs_household}
\title{Household Socio-Economic Classification from French individual PCS of the household member}
\usage{
gs_household(csind1, csind2)
}
\arguments{
\item{csind1}{2-digit PCS socio-economic classification categories of first household member}

\item{csind2}{2-digit PCS socio-economic classification categories of second household member (if any)}
}
\description{
Computes the synthetic socio-economic class of a household from the individual socio-economic classes of its members
}
\details{
The function computes the French synthetic socio-economic class of a household by following the rules indicated in CNIS, 2019, p. 42 sq. The resulting classification distinguishes 7 groups, coded with latin numerals):\tabular{ll}{
   Code \tab Category \cr
   I \tab Ménages à dominante cadre \cr
   II \tab Ménages à dominante intermédiaire ou cadre \cr
   III \tab Ménages à dominante employée ou intermédiaire \cr
   IV \tab Ménages à dominante petit indépendant \cr
   V \tab Ménages à dominante ouvrière \cr
   VI \tab Ménages mono-actifs d’employé ou d’ouvrier \cr
   VII \tab Ménages inactifs \cr
}
}
\seealso{
\code{\link[datatools]{gs_household}} for a detailed classification in 16 subgroups.
}
