% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs.R
\name{gsind_to_cshouse}
\alias{gsind_to_cshouse}
\title{Coding Individual GS to Household CS or GS}
\usage{
gsind_to_cshouse(gsind1, gsind2)
}
\arguments{
\item{gsind1}{2-digit PCS socio-economic classification categories of first household member}

\item{gsind2}{2-digit PCS socio-economic classification categories of second household member (if any)}
}
\description{
# Création de la variable csmenage à partir des 2 variables gsind à 1 chiffre de la personne de référence et du conjoint.
}
\details{
Cette fonction n'est pas utilisée directement, elle est mobilisée dans les fonctions cs_household() et gs_household()
}
