% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{hash}
\alias{hash}
\title{hash}
\usage{
hash(x, A = 0.618034, M = 1e+06, skip = TRUE, N = 11)
}
\arguments{
\item{x}{A character vector to hash}

\item{A}{Default value is (sqrt(5) - 1) / 2. See Donald Knuth, 'The Art of
Computer Programming', 1968}

\item{M}{Hash table size (limits the range of hash values)}

\item{skip}{If TRUE then pick every second byte (i.e. drop the UTF code page)}

\item{N}{Modular scaling for character codes}
}
\value{
A numeric vector of hash values
}
\description{
Simple hash function
}
\examples{

#

}
