% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{init project function}
\usage{
init(..., init_path = NULL)
}
\arguments{
\item{...}{}

\item{init_path}{give the path to the init file}
}
\description{
init project function
}
\details{
init project function
}
\author{
Jochen Kruppa
}
