% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_library}
\alias{install_library}
\title{Installation and Loading of Packages}
\usage{
install_library(packages)
}
\arguments{
\item{pkgs}{character vector of the names of packages to install (if they are not already installed) and load.}
}
\description{
Checks whether packages are installed before loading
}
\details{
The function first checks whether packages are installed, and downloads from the CRAN repository those which are not already present, before loading them by calling the library() function.
}
\examples{
install_library("datatools")
install_library(c("questionr", "explor")
}
\seealso{
\code{\link[utils]{install.packages}}, \code{\link[base]{library}}
}
