% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is.named}
\alias{is.named}
\title{Check for named vector}
\usage{
is.named(file)
}
\arguments{
\item{x}{}
}
\value{
Check for named vector
}
\description{
Check for named vector
}
\details{
Check for named vector
}
\author{
Jochen Kruppa
}
