% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender.R
\name{is_female}
\alias{is_female}
\title{Predict whether the genders of first names are male.}
\usage{
is_female(firstname, year_min = 1900, year_max = 2017)
}
\arguments{
\item{firstname}{First names as a character vector.}
}
\value{
Logical. TRUE if the input first name is female, FALSE otherwise. NA is returned when the first name is unknown in the database.
}
\description{
This function predicts whether the genders of first names are female.
}
\examples{
is_female(c("Baptiste", "Annick")
}
\seealso{
\code{\link{gender_unique}}, \code{\link{gender}}, \code{\link{is_male}}
}
