% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Join.R
\name{Join}
\alias{Join}
\title{Join Class}
\description{
Join Class

Join Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Join-new}{\code{Join$new()}}
\item \href{#method-Join-test}{\code{Join$test()}}
\item \href{#method-Join-print}{\code{Join$print()}}
\item \href{#method-Join-add}{\code{Join$add()}}
\item \href{#method-Join-add_all}{\code{Join$add_all()}}
\item \href{#method-Join-clone}{\code{Join$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Join-new"></a>}}
\if{latex}{\out{\hypertarget{method-Join-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Join$new(relationship)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{relationship}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Join-test"></a>}}
\if{latex}{\out{\hypertarget{method-Join-test}{}}}
\subsection{Method \code{test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Join$test()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Join-print"></a>}}
\if{latex}{\out{\hypertarget{method-Join-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Join$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Join-add"></a>}}
\if{latex}{\out{\hypertarget{method-Join-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Join$add(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Join-add_all"></a>}}
\if{latex}{\out{\hypertarget{method-Join-add_all}{}}}
\subsection{Method \code{add_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Join$add_all()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Join-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Join-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Join$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
