% Generated by roxygen2 (4.0.1): do not edit by hand
\name{keyMap.CB}
\alias{keyMap.CB}
\title{KeyMap function for CB project}
\usage{
keyMap.CB(keys, inputKeyType = "InnerCode", outputKeyType = "JTID",
  region = "CN", keyMapFile = NULL)
}
\arguments{
\item{keys}{keys to be mapped}

\item{inputKeyType}{}

\item{outputKeyType}{}

\item{region}{}

\item{keyMapFile}{}
}
\value{
vector
}
\description{
KeyMap function for CB project
}
\details{
KeyMap function for CB project
}
\author{
Weilin Lin
}

