% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill_format_tbl.R
\name{kill_format_tbl}
\alias{kill_format_tbl}
\title{Disables all formatting of the tibble() package}
\usage{
kill_format_tbl()
}
\description{
This function overrides the tibble print() function 
such that calling print() on a tbl object displays 
the first 10 rows in the format of a data.frame in base R. 
Call after loading the tibble package.
}
\examples{
\dontrun{

library(tibble)

tibble(x = runif(10), y = x * 2)

kill_format_tbl()
tibble(x = runif(10), y = x * 2)

} 



}
