% Generated by roxygen2 (4.0.1): do not edit by hand
\name{liquidity}
\alias{liquidity}
\title{Read liquidity data of bond}
\usage{
liquidity(startDate, endDate = startDate, jtids = NULL, vars = NULL,
  region = "CN", inDir = NULL, ...)
}
\arguments{
\item{startDate}{The startDate of Data}

\item{endDate}{The endDate of Data}

\item{jtids}{symbols}

\item{vars}{variables in data}

\item{region}{region for data}

\item{inDir}{file direction for writing in}

\item{...}{other params}
}
\value{
data frame
}
\description{
Read liquidity data of bond
}
\details{
Read liquidity data of bond
}
\author{
Weilin Lin
}

