% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{list2Biostring}
\alias{list2Biostring}
\title{Small function to return list to DNAStringSet}
\usage{
list2Biostring(list)
}
\arguments{
\item{list}{}
}
\value{
DNAStringSet
}
\description{
Small function to return list to DNAStringSet
}
\details{
Small function to return list to DNAStringSet
}
\author{
Jochen Kruppa
}
