% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{loadNamedFasta}
\alias{loadNamedFasta}
\title{Load named fasta file}
\usage{
loadNamedFasta(file)
}
\arguments{
\item{file}{file to be loaded}
}
\value{
DNAStringSet
}
\description{
Load named fasta file
}
\details{
You get the same result with readDNAStringSet, but this function
is used in some older projects and can not be dismissed
}
\author{
Jochen Kruppa
}
