% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bins.R
\name{make_bins}
\alias{make_bins}
\title{Discretize a variable either by user-defined breaks or by quantiles}
\usage{
make_bins(x, breaks = NULL, quantiles = NULL)
}
\arguments{
\item{x}{vector to discretize}

\item{breaks}{vector of breaks}

\item{quantiles}{vector of quantiles or number of evenly spaced quantiles}
}
\value{
integer vector
}
\description{
Discretize a variable either by user-defined breaks or by quantiles
}
\details{
If no breaks or quantiles are given two bins are formed with \code{mean(x)} as break.
}
\examples{
 \dontrun{

 require(dplyr)

 w <- sort(rnorm(10))
 make_bins(w,breaks=c(-1,0,1))

 } 



}
