% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_id.R
\name{make_id}
\alias{make_id}
\title{Creates a unique, numeric ID sequence based on a set of variables}
\usage{
make_id(...)
}
\arguments{
\item{...}{variables to be used to create unique ID-sequence}
}
\value{
vector that contains unique numeric id sequence starting with 0 up to K.
}
\description{
Creates a unique, numeric ID sequence based on a set of variables
}
\examples{
 \dontrun{

 require(dplyr)

 mtcars \%>\% mutate( id = make_id(cyl, vs) )

 } 



}
