% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_series_id}
\alias{make_series_id}
\title{Creates a unique ID for all uninterrupted series}
\usage{
make_series_id(x)
}
\arguments{
\item{x}{numeric vector to be used}
}
\value{
an integer vector
}
\description{
Creates a unique ID for all uninterrupted series
}
\details{
Creates a unique, numeric ID for each uninterrupted sequence of equal values in a vector
}
\examples{
 \dontrun{

 x <- c(0,1,1,1,2,2,1,3,1,0,0)

 make_series_id(x)

 } 



}
