% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTable.R
\name{makeTable}
\alias{makeTable}
\title{makeTable}
\usage{
makeTable(data, coreColumns, columns, number, padSuffix, namesResult,
  na.check = 1)
}
\arguments{
\item{data}{A data.frame containing data to be worked withs}

\item{coreColumns}{Metadata / identifying data}

\item{columns}{Columns (without suffixes) to be included}

\item{number}{Number of grouped columns (used to create column suffixes)}

\item{padSuffix}{Logical. Do we add suffixes with "_0n" rather than "_n".}

\item{namesResult}{Column names in result}

\item{na.check}{Column number to use for checking NA values; default is 1.}
}
\value{
A data.frame in long format.
}
\description{
Make separate tables in long format
}
\examples{

#

}
