% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{merge_calls}
\alias{merge_calls}
\title{Merge two calls.}
\usage{
merge_calls(c1, c2)
}
\arguments{
\item{c1}{First call.}

\item{c2}{Second call.}
}
\value{
A merged call using arguments from both c1 and c2.
}
\description{
Merge two calls.
}
