% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeByDate.R
\name{mergeByDate}
\alias{mergeByDate}
\title{Merge by Date}
\usage{
mergeByDate(ds, ds2add)
}
\arguments{
\item{ds}{parent data.frame to be added to and return}

\item{ds2add}{data to pull covariates from (assumes 3 columns: SID, DATE, VAR)}
}
\description{
mergeByDate function adds new column to ds that contains
 VAR from nearest past DATE for SID to each row
 and returns ds
}
\examples{
mergeByDate()
}
\keyword{dataTools}
