% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mk_dir}
\alias{mk_dir}
\title{Small wrapper for dir.create}
\usage{
mk_dir(dir_path)
}
\arguments{
\item{dir_path}{}
}
\description{
Small wrapper for dir.create
}
\details{
Small wrapper for dir.create
}
\author{
Jochen Kruppa
}
