% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mktData.CB}
\alias{mktData.CB}
\title{Read data from our server}
\usage{
mktData.CB(startDate, endDate = startDate, jtids = NULL, vars = NULL,
  region = "CN", inDir = NULL, ...)
}
\arguments{
\item{startDate}{The startDate of Data}

\item{endDate}{The endDate of Data}

\item{jtids}{symbols}

\item{vars}{variables in data}

\item{region}{region for data}

\item{inDir}{file direction for writing in}

\item{...}{other parametres}
}
\value{
An OHLC files
}
\description{
Read data from our server
}
\details{
Read market data of A share and HK market stock data from server
}
\author{
Weilin Lin
}

