% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mktData}
\alias{mktData}
\title{Read data from our server}
\usage{
mktData(startDate, endDate = startDate, jtids = NULL, vars = NULL,
  region = "CN", inDir = NULL, source = "MySQL", ...)
}
\arguments{
\item{startDate}{The startDate of Data}

\item{endDate}{The endDate of Data}

\item{jtids}{symbols}

\item{vars}{variables in data}

\item{region}{region for data}

\item{inDir}{file direction for writing in}

\item{source}{}

\item{...}{other parametres}
}
\value{
An OHLC files
}
\description{
Read data from our server
}
\details{
Read market data of A share and HK market stock data from server
}
\author{
Weilin Lin
}

