% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mktDataHFT.CB}
\alias{mktDataHFT.CB}
\title{read HFT data from database}
\usage{
mktDataHFT.CB(startDate, endDate = startDate, jtids = NULL, vars = NULL)
}
\arguments{
\item{startDate}{}

\item{endDate}{}

\item{jtids}{}

\item{vars}{}
}
\value{
data frame
}
\description{
read HFT data from database
}
\details{
read HFT data from database
}
\author{
Weilin Lin
}

