% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mktDataHFT}
\alias{mktDataHFT}
\title{read HFT data from database}
\usage{
mktDataHFT(startDate, endDate = startDate, securityID = NULL, vars = NULL)
}
\arguments{
\item{startDate}{}

\item{endDate}{}

\item{securityID}{}

\item{vars}{}
}
\value{
data frame
}
\description{
read HFT data from database
}
\details{
read HFT data from database
}
\author{
Weilin Lin
}

