% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mktStat.CTDBond}
\alias{mktStat.CTDBond}
\title{Statistics of the deliverable bonds}
\usage{
mktStat.CTDBond(startDate, endDate = startDate, contract = NULL,
  region = "CN", inDir = NULL, jtids = NULL, vars = NULL,
  onlyCTD = FALSE, ...)
}
\arguments{
\item{startDate}{startDate of the data}

\item{endDate}{endDate of the data}

\item{contract}{specify the contract}

\item{region}{equal to CN}

\item{inDir}{}

\item{jtids}{}

\item{vars}{}

\item{onlyCTD}{}

\item{...}{}
}
\value{
data frame
}
\description{
Statistics of the deliverable bonds
}
\details{
Statistics of the deliverable bonds
}
\author{
Weilin Lin
}

