% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mktStat}
\alias{mktStat}
\title{Read marker statistics from our database}
\usage{
mktStat(startDate, endDate = startDate, jtids = NULL, vars = NULL,
  region = "CN", inDir = NULL, ...)
}
\arguments{
\item{startDate}{start date}

\item{endDate}{end date}

\item{jtids}{symbols of stocks or coupon}

\item{vars}{variables in data}

\item{region}{region for stocks or coupon, must equal to "CN" or "HK"}

\item{inDir}{direction, unused parameter}

\item{...}{other parameters}
}
\value{
data frame with liquidity information and others
}
\description{
read marker statistics from our database
}
\details{
read marker statistics from our database
}
\author{
Weilin Lin
}

