% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mktTF}
\alias{mktTF}
\title{This is the script to make trearury future market data.}
\usage{
mktTF(startDate, endDate = startDate, jtids = NULL, vars = NULL,
  region = "CN", inDir = NULL, ...)
}
\arguments{
\item{startDate}{start Date}

\item{endDate}{end Date}

\item{jtids}{symbols of stocks or coupon}

\item{vars}{varibles in data}

\item{region}{region for stocks or coupon}

\item{inDir}{direction}

\item{...}{other parameters}
}
\value{
null
}
\description{
This is the script to make trearury future market data.
}
\details{
This is the script to make trearury future market data.
}
\author{
Weilin Lin
}

