% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mktValuation}
\alias{mktValuation}
\title{read market valuation from our database}
\usage{
mktValuation(startDate, endDate = startDate, jtids = NULL, vars = NULL,
  region = "CN", inDir = NULL, ...)
}
\arguments{
\item{startDate}{start Date}

\item{endDate}{end Date}

\item{jtids}{symbols of stocks or coupon}

\item{vars}{variables in data}

\item{region}{region for stocks or coupon}

\item{inDir}{Direction}

\item{...}{other parameters}
}
\value{
null
}
\description{
read market valuation from our database
}
\details{
read market valuation from our database
}
\author{
Weilin Lin
}

