% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandling.r
\name{modal}
\alias{modal}
\title{Get the modal value of vector}
\usage{
modal(vector)
}
\arguments{
\item{vector}{}
}
\value{
modal value
}
\description{
Get the modal value of vector
}
\details{
Get the modal value of vector
}
\author{
Jochen Kruppa
}
