% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericTime.R
\name{numericTime}
\alias{numericTime}
\title{Convert Clock Time to Numeric Time}
\usage{
numericTime(vec)
}
\arguments{
\item{vec}{Vector containing time values of format "HH:MM" stored as
\code{character}.}
}
\value{
Vector containing time values converted to fractional hour +
  minute/60 as \code{numeric}.
}
\description{
Clock times stored as character are converted into numeric hours for ease of
calculating time differences.
}
\details{
This function was created to convert clock times to numeric times with
minutes converted to fractions of hours.
}
\examples{
clockTimes <- c("10:06", "8:30", "12:15")
numericTime(clockTimes)

}
\author{
Samuel Callisto \email{calli055@umn.edu}
numericTime()
}
