% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols.R
\name{ols}
\alias{ols}
\title{This function is a wrapper around \code{lm} and \code{coeftest}}
\usage{
ols(..., HC = "const", cluster = NULL, data_frame = FALSE, as_attr = TRUE)
}
\arguments{
\item{...}{(required) arguments passed to \code{lm}}

\item{HC}{the HC matrix to be used for the robust standard error calculations}

\item{cluster}{the name of the cluster variable in the \code{data.frame} passed to \code{lm} 
for the cluster robust standard error calculations}

\item{data_frame}{return estimates in a \code{data.frame} or \code{coeftest} object(default)}

\item{as_attr}{return \eqn{R^2} and \eqn{N} as column or attributes when \code{data_frame=TRUE}?}
}
\value{
a \code{data.frame} if \code{data_frame=TRUE} otherwise a \code{coeftest} object. 
The \code{data.frame} object has two additional attributes "N" (the number of observations) and 
"r2" the \eqn{R^2} of the model.
}
\description{
This function is a wrapper around \code{lm} and \code{coeftest}
}
\details{
This function calls \code{lm()} and passes the model object further to \code{coeftest()} to 
calculate (cluster robust) standard errors. The results are equivalent to calling 
\code{coeftest(model, vcov=vcovHC(m,HC))} or \code{cl(model,cluster)} after estimating
\code{model} via \code{lm()}.

The default settings produce estimates that are identical to calling \code{lm(...)}.

To reproduce STATA's default robust standard errors use \code{HC="HC1"}, to reproduce the 
STATA cluster robust standard errors (\code{, vce(cluster ___)}), supply the name of the cluster variable.
}
\examples{

x <- runif(100, -1, 1)
 g <- sample(1:10, 100, replace=TRUE)
 y <- 1 + x + rnorm(100)

 dat <- data.frame(y=y,x=x,g=g)
 dat[1,1] <- NA; dat[10,1] <- NA

 ols(y~x,dat=dat,cluster='g', na.action='na.omit')


}
\seealso{
\code{\link{lm}}, \code{\link{vcovHC}}
}
\author{
The cluster robust standard error function \code{cl()} by Mahmood Ara at Stockholm University, see:
	\url{https://thetarzan.wordpress.com/2011/06/11/clustered-standard-errors-in-r/}
}
