% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered_factor.R
\name{ordered_factor}
\alias{ordered_factor}
\title{Creates an ordered factor based on order of another variable}
\usage{
ordered_factor(var, seq)
}
\arguments{
\item{var}{variable to be converted into a factor}

\item{seq}{variable that implies the sequence of \code{var} when the function \code{order} is applied to it}
}
\value{
\code{factor(var)} with the implied order by \code{seq}.
}
\description{
Creates an ordered factor based on order of another variable
}
\details{
This function helps to re-order groups in a \code{ggplot2} plot. See example below.
}
\examples{
 \dontrun{

 require(ggplot2)
 require(dplyr)

 economics_long <- economics_long \%>\% group_by(variable) \%>\% 
	 mutate( value01_mean=mean(value01) ) 

	economics_long <- ungroup(economics_long) \%>\% 
	 mutate(variable_mean_ordered = ordered_factor(as.character(variable),value01_mean) )

 ggplot(economics_long, aes(date, value01, colour = variable)) +
 	geom_line()

 ggplot(economics_long, aes(date, value01, colour = variable_mean_ordered)) +
 	geom_line()

 }

}
