% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin.R
\name{origin}
\alias{origin}
\title{Get origins of first names}
\usage{
origin(firstname)
}
\arguments{
\item{firstname}{first names as a character string.}
}
\value{
The geographic origins of the first names. NA is returned when the input first names are unknown in the database.
}
\description{
This function returns the geographic origins of first names.
}
\examples{
origin("Baptiste", "Martha", "Lars", "Ahmed")
}
