% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin.R
\name{origin_unique}
\alias{origin_unique}
\title{Get origin of first name}
\usage{
origin_unique(first_name)
}
\arguments{
\item{first_name}{first name as a character string.}
}
\value{
The geographic origin of the first name. NA is returned when the input first name is unknown in the database.
}
\description{
This function returns the geographic origin of a first name.
}
\examples{
origin_unique("Baptiste")
origin_unique("Steven")
}
