% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package}
\alias{package}
\title{R internal function to build and compile packages}
\usage{
package(packageName, method = "compile", version = "minor",
  pkgDir = "C:/Users/157216/source/r-pkg")
}
\arguments{
\item{packageName}{Name of the package}

\item{method}{'build' or 'compile' [default]}

\item{version}{'minor' [default] or 'major', the version number
will be incremented at the last position (minor) or second last
(major).}
}
\description{
Small wrapper function
}
\details{
Small wrapper function
}
\examples{
library(devtools)

## Step 1: Building the package

package("kmerPyramid", method = "build")

## Step 2: Compile the package

package("kmerPyramid", method = "compile", version = "minor")

## [optional] add data to packages

devtools::use_data(par, pkg = "dataTools", overwrite = TRUE)
}
\author{
Jochen Kruppa
}
