% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_func.R
\name{paste_func}
\alias{paste_func}
\title{Paste Function}
\usage{
paste_func(df, cols, string)
}
\arguments{
\item{df}{A DataFrame}

\item{cols}{A vector of column numbers to paste the string onto}

\item{string}{The string to paste onto the beginning of each specified column name}
}
\description{
Given a dataframe, a vector of column numbers, and a string, this adds the string onto each defined column name.
Returns dataframe with pasted names.
}
\examples{
paste_func(iris, c(1:ncol(iris)), "PASTETHIS_")
}
\keyword{dataframe}
