% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print.DataFrame}
\alias{print.DataFrame}
\title{S3 print method of \code{DataFrame}}
\usage{
\method{print}{DataFrame}(x, ...)
}
\arguments{
\item{x}{A \code{DataFrame} object.}

\item{...}{Not used.}
}
\value{
Invisibly returns the \code{DataFrame}.
}
\description{
S3 print method of \code{DataFrame}
}
