% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_initial_vals.R
\name{print_initial_vals}
\alias{print_initial_vals}
\title{A Cat Function}
\usage{
print_initial_vals(df, x)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
