% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q.R
\name{Q}
\alias{Q}
\title{Function to discretize a variable based on a selected quantile}
\usage{
Q(x, q)
}
\arguments{
\item{x}{variable to discretize}

\item{q}{probability corresponding to quantile}
}
\value{
\code{vector} with a discretized version of x.
}
\description{
Function to discretize a variable based on a selected quantile
}
\examples{
 \dontrun{

 x <- runif(100,-1,1)
 y <- x^2 + rnorm(100)
 
 summary( lm(y ~ Q(x,0.25) ) )
 summary( lm(y ~ Q(x,0.75) ) )
 
 }

}
