% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmode.R
\name{qmode}
\alias{qmode}
\title{Value occuring most often (simple mode)}
\usage{
qmode(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
scalar estimate.
}
\description{
Value occuring most often (simple mode)
}
\details{
If NA is mode, then qmode(x)=NA. 
If there are two modes, qmode(x) takes the first.
}
\examples{
 \dontrun{

 qmode(rpois(10, 1))
 
 }


}
