% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ods_sheets.R
\name{read_ods_sheets}
\alias{read_ods_sheets}
\title{datatools::read_ods_sheets}
\usage{
read_ods_sheets(path, skip = 0)
}
\arguments{
\item{path}{path to the ods file.}

\item{skip}{An integer: number of rows to be skiped in the beginning of each sheet; default = 0}
}
\value{
A series of tibbles corresponding to the different sheets contained in the Excel workbook.
}
\description{
Import LibreOffice workbook with multiple sheets
}
\details{
This function imports multiple sheets contained in a .ods spreadsheet into distinct tibbles.
}
\examples{
\dontrun{
path <- ".../myworkbook.ods"
read_ods_sheets(path)
read_ods_sheets(path, skip = 2)
}
}
