% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xl_sheets.R
\name{read_xl_sheets}
\alias{read_xl_sheets}
\title{datatools::read_xl_sheets}
\usage{
read_xl_sheets(path, skip)
}
\arguments{
\item{path}{path to the Excel workbook .}

\item{skip}{An integer: number of rows to be skiped in the beginning of each sheet.}
}
\value{
A series of tibbles corresponding to the different sheets contained in the Excel workbook.
}
\description{
Import Excel workbook with multiple sheets
}
\details{
This function imports multiple Excel sheets into tibbles in R Global Environment.
}
\examples{
\dontrun{
path <- ".../myworkbook.xls"
read_xl_sheets(path)
read_xl_sheets(path, skip = 2)
}
}
