% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{Rel}
\alias{Rel}
\title{Relationship constructor sugar}
\usage{
Rel(left, right)
}
\arguments{
\item{left}{table}

\item{right}{table}
}
\value{
A \code{DataFrame} object.
}
\description{
Relationship constructor sugar
}
