% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Relationship.R, R/test.R
\name{Relationship}
\alias{Relationship}
\alias{Rel}
\title{Relationship class}
\usage{
Rel(left, right)
}
\description{
Relationship class

Relationship class

Title
}
\section{Functions}{
\itemize{
\item \code{Relationship}: Object

\item \code{Rel()}: Wrapper

}}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Relationship-new}{\code{Relationship$new()}}
\item \href{#method-Relationship-get_on}{\code{Relationship$get_on()}}
\item \href{#method-Relationship-on}{\code{Relationship$on()}}
\item \href{#method-Relationship-is_not_fully_specified}{\code{Relationship$is_not_fully_specified()}}
\item \href{#method-Relationship-clone}{\code{Relationship$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-new"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$new(left, right)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{right}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-get_on"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-get_on}{}}}
\subsection{Method \code{get_on()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$get_on()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-on"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-on}{}}}
\subsection{Method \code{on()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$on(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-is_not_fully_specified"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-is_not_fully_specified}{}}}
\subsection{Method \code{is_not_fully_specified()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$is_not_fully_specified()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
