% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem_sparse_cat.R
\name{rem_sparse_cat}
\alias{rem_sparse_cat}
\title{Remove sparse categories}
\usage{
rem_sparse_cat(x, threshold, label = "(Other)")
}
\arguments{
\item{x}{factor variable}

\item{threshold}{threshold}

\item{label}{label for the new category}
}
\value{
factor variable
}
\description{
Remove sparse categories
}
\details{
Combines levels with strictly less than \code{threshold} number of observations into a new category. NA are ignored.
}
\examples{

 a <- rpois(100,1)
 k <- length(unique(a))
 a <- factor(a, labels=LETTERS[1:k])
 rem_sparse_cat(a, threshold=20)


}
