% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_cols.R
\name{remove_cols}
\alias{remove_cols}
\title{Remove Cols Function}
\usage{
remove_cols(df, x)
}
\arguments{
\item{df}{A DataFrame}

\item{x}{Number of non-na items to print for each column.}
}
\description{
Given a dataframe and a number x, prints the first x non-na values of each column (though prints first 5 values if no number "x" is specified).
}
\examples{
print_initial_vals(iris)
}
\keyword{dataframe}
