% Generated by roxygen2 (4.0.1): do not edit by hand
\name{report.dailySummary}
\alias{report.dailySummary}
\title{report summary for contracts}
\usage{
report.dailySummary(date)
}
\arguments{
\item{date}{}
}
\value{
data frame
}
\description{
report summary for contracts
}
\details{
report summary for contracts
}
\author{
Weilin Lin
}

