% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retype.R
\name{retype}
\alias{retype}
\title{Retype Variable}
\usage{
retype(df, ...)
}
\arguments{
\item{df:}{data.frame with values to typecast}

\item{...:}{unquoted list alternating variables and datatypes to convert to}
}
\value{
Data.frame \code{ds} with variables listed typecast to
}
\description{
retype is a typecasting function which can be used
in the tidy workflow to change the type of multiple
functions within a data.frame in-line.
}
\examples{
iris2 <- iris \%>\% select(Species, Sepal.Length) \%>\% retype(Species, character)
str(iris2)

}
\author{
Samuel Callisto \email{calli055@umn.edu} retype()
}
