% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandling.r
\name{round0}
\alias{round0}
\title{A p value rounder}
\usage{
round0(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector
}
\description{
A p value rounder
}
\details{
A p value rounder
}
\author{
Jochen Kruppa
}

