% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{runCMD}
\alias{runCMD}
\title{Lazy wrapper for system}
\usage{
runCMD(cmd, intern = FALSE)
}
\arguments{
\item{cmd}{pasted command to run}
}
\description{
Lazy wrapper for system
}
\details{
Lazy wrapper for system
}
\author{
Jochen Kruppa
}
