% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runKnitr.r
\name{runKnitr}
\alias{runKnitr}
\title{Compiler for RNW and TEX files in subdirs}
\usage{
runKnitr(files, type = NULL, copy.dir = NULL, force = FALSE)
}
\arguments{
\item{files}{RNW or TEX files}

\item{type}{determines the type of files. Default NULL self extracting}

\item{copy.dir}{where should the final pdfs copy to?}
}
\description{
Compiler for RNW and TEX files in subdirs
}
\details{
Compiler for RNW and TEX files in subdirs
}
\author{
Jochen Kruppa
}

