% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastx.R
\name{sam2bam}
\alias{sam2bam}
\title{Finds sam files and convert to bam}
\usage{
sam2bam(samDir, pattern, keep = TRUE)
}
\arguments{
\item{samDir}{Dir where are 'pattern' sam's should be converted}

\item{pattern}{Pattern to find sam files to be converted}
}
\description{
Finds sam files and convert to bam
}
\details{
Finds sam files and convert to bam
}
\author{
Jochen Kruppa
}

