% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_base.R
\name{set_base}
\alias{set_base}
\title{Rescales vector to base value}
\usage{
set_base(v, base)
}
\arguments{
\item{base}{value that will equal 1}

\item{x}{numeric vector}
}
\value{
standardized \code{vector}.
}
\description{
Rescales vector to base value
}
\details{
The function rescales all values such that the value \code{base} in \code{x} equals 1 in the returned vector.
}
\examples{
 \dontrun{

 x <- sort(rpois(10, 5))
 tr(x,3)
 
 }


}
