% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_log.R
\name{source_log}
\alias{source_log}
\title{Alternative to \code{source()} that also logs the output via \code{sink()}}
\usage{
source_log(file, ...)
}
\arguments{
\item{file}{file name}

\item{...}{parameters passed to \code{\link{source}}}
}
\description{
Alternative to \code{source()} that also logs the output via \code{sink()}
}
\details{
The logfile is written to ./logs/ using \code{file}
as the filename appended with a timestamp.

The header and footer are appended with a timestamp 
and the footer features the output of \code{\link{sessionInfo}}.
}
\seealso{
\code{\link{source}} and \code{\link{sink}}
}
