% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize2}
\alias{standardize2}
\alias{standardize}
\title{Standardize a Vector}
\usage{
standardize2(x, scale = TRUE)

standardize(x, scale = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{scale}{scale variable by (twice) its standard deviation?}
}
\value{
standardized \code{vector}.
}
\description{
Standardize a Vector
}
\details{
\code{standardize} centers values and scales them by their standard deviation. \code{standardize2} does the same but scales by twice their standard deviation.
}
\examples{

 x <- runif(100, 0, 1.2)
 xz <- standardize2(x)
 sd(xz)
 
 xz <- standardize(x)
 sd(xz)

}
\references{
Gelman, A. (2008). Scaling Regression Inputs by Dividing by two Standard Deviations. Statistics in Medicine 27(2865–2873).
}
