% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{stringify_dots}
\alias{stringify_dots}
\title{Return ellipsis as character vector.}
\usage{
stringify_dots(...)
}
\arguments{
\item{...}{Arbitrary values.}
}
\value{
A character vector with quoted arguments.
}
\description{
Return ellipsis as character vector.
}
