% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiHo.r
\name{time_worked}
\alias{time_worked}
\title{Small function to calculate the working time (improved)}
\usage{
time_worked(start, lunch_start, lunch_end, end, workaccount)
}
\arguments{
\item{end}{}

\item{workaccount}{}

\item{from}{}

\item{lunchStart}{}

\item{lunchEnd}{}
}
\value{
data.frame
}
\author{
Jochen Kruppa
}
