% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeAfterDose.R
\name{timeAfterDose}
\alias{timeAfterDose}
\title{Convert Clock Time to Time After Dose}
\usage{
timeAfterDose(dataset, ID, TIME, DATE)
}
\arguments{
\item{ID}{the column number contianing ID in ds}

\item{TIME}{the column number containing TIME in ds}

\item{DATE}{the column number containing DATE in ds}
}
\value{
vector containing TAD values
}
\description{
Short description.
}
\details{
Long description.
}
\examples{



}
\author{
Samuel Callisto \email{calli055@umn.edu}
}
