% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiHo.r
\name{timeWorked}
\alias{timeWorked}
\title{Small function to calculate the working time}
\usage{
timeWorked(from = NULL, lunchStart = NULL, lunchEnd = NULL, end = NULL,
  workaccount = NULL)
}
\arguments{
\item{from}{}

\item{lunchStart}{}

\item{lunchEnd}{}

\item{end}{}

\item{workaccount}{}
}
\value{
data.frame
}
\author{
Jochen Kruppa
}

