% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_plain.R
\name{to_plain}
\alias{to_plain}
\title{Convert UTF8 string to ASCII string}
\usage{
to_plain(s)
}
\arguments{
\item{x}{character vector}
}
\value{
\code{vector} with ASCII characters only
}
\description{
Convert UTF8 string to ASCII string
}
\details{
This function substitutes ASCII characters for (many) none-ASCII characters, e.g. ä,ö,ü
}
\examples{
 \dontrun{

 x <- "Viele Grüße nach Eisenhüttenstadt"
 to_plain(x)
 
 }

}
\references{
Source: \url{https://stackoverflow.com/questions/17517319/r-replacing-foreign-characters-in-a-string}
}
