% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topcode.R
\name{topcode}
\alias{topcode}
\title{Top-coding values in a vector}
\usage{
topcode(x, probs = 0.99)
}
\arguments{
\item{x}{numeric vector}

\item{probs}{scalar of the quantile cut-point.}
}
\value{
\code{vector} top-coded vector.
}
\description{
Top-coding values in a vector
}
\details{
This function sets all values larger than the selected quantile to the largest value of the remaining values.
}
\examples{
 \dontrun{

 x <- sort(rnorm(10))
 topcode(x)
 
 }

}
