% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender.R
\name{unaccent}
\alias{unaccent}
\title{Remove accents}
\usage{
unaccent(string)
}
\arguments{
\item{string}{input character vector.}
}
\value{
An unaccented character vector.
}
\description{
This function removes accents from character strings.
}
\examples{
unaccent("Jérémie")
unaccent("Héloïse")
}
