% Generated by roxygen2 (4.0.1): do not edit by hand
\name{univ}
\alias{univ}
\title{univ}
\usage{
univ(startDate, endDate = startDate, univType = "liquidityBased",
  univRange = c(0, 1), region, inDir = NULL, daysUsedToSmooth = 10, ...)
}
\arguments{
\item{startDate}{}

\item{endDate}{}

\item{univType}{}

\item{univRange}{}

\item{region}{}

\item{inDir}{}

\item{daysUsedToSmooth}{}

\item{...}{}
}
\value{
data frame
}
\description{
universe
}
\details{
universe
}
\author{
Weilin Lin
}

