% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_analysis.R
\name{use_analysis}
\alias{use_analysis}
\title{use_analysis}
\usage{
use_analysis(project_name = ".")
}
\arguments{
\item{project_name}{character; defaults to current directory}
}
\description{
Creates a project skeleton for data analysis.
}
\examples{
# Example


try({
  datatools::use_analysis(tempdir())
})

}
\author{
Reinhard Simon
}

