% Generated by roxygen2 (4.0.1): do not edit by hand
\name{volatility}
\alias{volatility}
\title{volatility of stocks}
\usage{
volatility(startDate, endDate = startDate, jtids = NULL, vars = NULL,
  region = "CN", inDir = NULL, ...)
}
\arguments{
\item{startDate}{}

\item{endDate}{}

\item{jtids}{}

\item{vars}{}

\item{region}{}

\item{inDir}{}

\item{...}{}
}
\value{
data framme
}
\description{
volatility of stocks
}
\details{
volatility of stocks
}
\author{
Weilin Lin
}

