% Generated by roxygen2 (4.0.1): do not edit by hand
\name{windUpdateJTID}
\alias{windUpdateJTID}
\title{get windIDs need to update with wind}
\usage{
windUpdateJTID()
}
\value{
vector
}
\description{
get windIDs need to update with wind
}
\details{
get windIDs need to update with wind
}
\author{
Weilin Lin
}

