% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIO.r
\name{write.table0}
\alias{write.table0}
\title{Small wrapper for write.table}
\usage{
write.table0(...)
}
\arguments{
\item{...}{}
}
\description{
Small wrapper for write.table
}
\details{
Small wrapper for write.table with the options col.names = FALSE, row.names = FALSE, quote = FALSE
}
\author{
Jochen Kruppa
}

