% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIO.r
\name{write_xlsx}
\alias{write_xlsx}
\title{This is a link to a older function}
\usage{
write_xlsx(list, file, oneFile = TRUE)
}
\arguments{
\item{link}{to saveList2Xls}
}
\description{
This is a link to a older function
}
