% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIO.r
\name{writeRDS}
\alias{writeRDS}
\title{SaveRDS wrapper}
\usage{
writeRDS(object, dir, verbose = TRUE)
}
\arguments{
\item{object}{which should be saved}

\item{dir}{where should the object be saved}

\item{verbose}{should the function talk?}
}
\description{
A writeRDS wrapper for more convienence
}
\details{
Caution this is a wrapper...
}
\author{
Jochen Kruppa
}

