% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIO.r
\name{writeXLS}
\alias{writeXLS}
\title{This is a link to a older function}
\usage{
writeXLS(list, file, oneFile = TRUE)
}
\arguments{
\item{link}{to saveList2Xls}
}
\description{
This is a link to a older function
}

