% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender.R
\name{year}
\alias{year}
\title{Predict years of birth from first names}
\usage{
year(firstname, year_min = 1946, year_max = 2017)
}
\arguments{
\item{year_min}{starting year of the period over which the prediction is computed.}

\item{year_max}{ending year of the period over which the prediction is computed.}
}
\value{
The predicted years of birth based on the proportions of individuals with the input first names born each year. The function returns the modes of the distributions, i.e. the years with the highest numbers of individuals born by the input firstnames. NAs are returned when the input first names are unknown in the database.
}
\description{
This function predicts the years of birth of individuals from a vector of first names.
}
\examples{
year(c("Baptiste", "Henriette")
}
