% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender.R
\name{year_unique}
\alias{year_unique}
\title{Predict year from first name}
\usage{
year_unique(first_name, year_min = 1946, year_max = 2017)
}
\arguments{
\item{first_name}{first name as a character string.}

\item{year_min}{starting year of the period over which the prediction is computed.}

\item{year_max}{ending year of the period over which the prediction is computed.}
}
\value{
The predicted year of birth based on the proportions of individuals with the input first name born each year. The function returns the mode of the distribution, i.e. the year with the highest number of individuals born by the input firstname. NA is returned when the input first name is unknown in the database.
}
\description{
This function predicts the year of birth from a first name.
}
\examples{
year("Baptiste")
year("Henriette")
}
