% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.data.frame.DataFrame}
\alias{as.data.frame.DataFrame}
\title{Convert \code{DataFrame} to \code{data.frame}}
\usage{
\method{as.data.frame}{DataFrame}(x, row.names = NULL, optional = NULL, ...)
}
\arguments{
\item{x}{A \code{DataFrame} object.}

\item{row.names}{Not used.}

\item{optional}{Not used.}

\item{...}{Not used.}
}
\value{
A \code{data.frame} object.
}
\description{
Convert \code{DataFrame} to \code{data.frame}
}
