% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrame.R, R/sugar.R
\name{DataFrame}
\alias{DataFrame}
\alias{DF}
\title{DataFrame Class}
\usage{
DF(tbl, key = NULL, id = NULL)
}
\arguments{
\item{tbl}{An object of class \code{data.frame}.}

\item{key}{Optional vector of column names. Setting a key sorts the table in RAM using the values of the key column(s). See Details.}

\item{id}{Optional ID of the object. Not currently used.}
}
\value{
A \code{DataFrame} object.
}
\description{
A class for tabular data objects of class \code{data.frame}.
}
\details{
The key parameter is passed to \code{data.table::setkey(key)}. If the parameter is not passed, the existing keys of \code{x} (if any) will be kept.
}
\section{Functions}{
\itemize{
\item \code{DF()}: DataFrame$new() constructor sugar

}}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{Data getter.}

\item{\code{columns}}{Column names object.}

\item{\code{key}}{Key getter and setter.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataFrame-new}{\code{DataFrame$new()}}
\item \href{#method-DataFrame-print}{\code{DataFrame$print()}}
\item \href{#method-DataFrame-sort}{\code{DataFrame$sort()}}
\item \href{#method-DataFrame-is_key_unique}{\code{DataFrame$is_key_unique()}}
\item \href{#method-DataFrame-count}{\code{DataFrame$count()}}
\item \href{#method-DataFrame-update_join}{\code{DataFrame$update_join()}}
\item \href{#method-DataFrame-left_join}{\code{DataFrame$left_join()}}
\item \href{#method-DataFrame-append}{\code{DataFrame$append()}}
\item \href{#method-DataFrame-apply}{\code{DataFrame$apply()}}
\item \href{#method-DataFrame-update}{\code{DataFrame$update()}}
\item \href{#method-DataFrame-add}{\code{DataFrame$add()}}
\item \href{#method-DataFrame-drop}{\code{DataFrame$drop()}}
\item \href{#method-DataFrame-remove}{\code{DataFrame$remove()}}
\item \href{#method-DataFrame-deep_clone}{\code{DataFrame$deep_clone()}}
\item \href{#method-DataFrame-clone}{\code{DataFrame$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$new(tbl, key = NULL, id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-print"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-sort"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-sort}{}}}
\subsection{Method \code{sort()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$sort(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-is_key_unique"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-is_key_unique}{}}}
\subsection{Method \code{is_key_unique()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$is_key_unique()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-count"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-count}{}}}
\subsection{Method \code{count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$count(by = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-update_join"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-update_join}{}}}
\subsection{Method \code{update_join()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$update_join(relationship, columns = NULL, where = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-left_join"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-left_join}{}}}
\subsection{Method \code{left_join()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$left_join(relationship, add = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-append"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-append}{}}}
\subsection{Method \code{append()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$append(..., fill = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-apply"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-apply}{}}}
\subsection{Method \code{apply()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$apply(columns, mapper, where = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-update"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$update(column, mapper, where = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-add"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$add(column, mapper, where = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-drop"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-drop}{}}}
\subsection{Method \code{drop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$drop(columns)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-remove"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-remove}{}}}
\subsection{Method \code{remove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$remove(where)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-deep_clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-deep_clone}{}}}
\subsection{Method \code{deep_clone()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$deep_clone()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
