% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools-package.R
\docType{package}
\name{datatools-package}
\alias{datatools-package}
\alias{datatools}
\title{Simple tools I tend to use a lot when working with data in R}
\description{
Feedback is very welcome!
}
\details{
\tabular{ll}{
	Package: \tab datatools\cr
	Type: \tab Package\cr
	Version: \tab 0.0.0.9000\cr
	Date: \tab 2017-02-11\cr
	License: \tab  GPL-3\cr
	}
}
\author{
Moritz Marbach \email{marbach@ipz.uzh.ch}
}
