% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\docType{package}
\name{datatools}
\alias{datatools}
\alias{datatools-package}
\title{datatools: A package for summarising and processing data}
\description{
Data processing mostly for fisheries applications
}
