% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPatients.R
\name{extractPatients}
\alias{extractPatients}
\title{Extract Patients}
\usage{
extractPatients(ds, px)
}
\arguments{
\item{ds}{parent data.frame to extract data from}

\item{px}{list of patient identifiers to extract from the dataset}
}
\description{
This function extracts a subset of patient data from a larger dataset
}
\author{
Samuel Callisto \email{calli055@umn.edu}
}
\keyword{dataTools}
