% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs.R
\name{gs_household}
\alias{gs_household}
\title{Household Socio-Economic Classification from French individual PCS of the household member}
\usage{
gs_household(csind1, csind2)
}
\arguments{
\item{csind1}{2-digit PCS socio-economic classification categories of first household member}

\item{csind2}{2-digit PCS socio-economic classification categories of second household member (if any)}
}
\description{
Computes the synthetic socio-economic class of a household from the individual socio-economic classes of its members
}
\details{
# Création de la variable gemenage à un chiffre romain, à partir des 2 variables individuelles csindiv à 2 chiffres de la personne de référence et du conjoint. Cette fonction est utilisée directement pour créer la PCS Ménage.
}
\seealso{
\code{\link[datatools]{gs_household}}
}
