% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headr.R
\name{headr}
\alias{headr}
\title{headr()}
\usage{
headr(file, header.row = 1, data.start = 3)
}
\arguments{
\item{file}{location of .csv containing dataset with header followed by single unit row}
}
\description{
This function reads in a dataset with a header and a unit row
ignoring the unit row in preserving data types
}
\author{
Samuel Callisto \email{calli055@umn.edu}
headr()
}
\keyword{dataTools}
