% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hungarian_merge.R
\name{hungarian_merge}
\alias{hungarian_merge}
\title{Merge Two Data Frames using the Hungarian Method}
\usage{
hungarian_merge(x, y, by.x = NULL, by.y = NULL)
}
\arguments{
\item{x, y}{data frames with the same number of observations, or objects to be coerced to one.}

\item{by.x, by.y}{specifications of the columns used for merging.}
}
\description{
Merge two data frames by a common column, s.t. the Damerau-Levenshtein distance 
is minimized
}
\details{
Finds the optimal matches using (a fast version of) the Hungarian method as implemented in 
\code{\link[adagio]{assignment}}.
}
\seealso{
\code{\link[adagio]{assignment}} and \code{\link[stringdist]{stringdist}}
}
