% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_series_id}
\alias{make_series_id}
\title{Creates a unique ID for all uninterrupted series in a vector}
\usage{
make_series_id(x, delta = 0)
}
\arguments{
\item{x}{numeric vector to be used}

\item{delta}{increment with default \code{delta}=0}
}
\value{
an integer vector
}
\description{
Creates a unique ID for all uninterrupted series in a vector
}
\details{
If x[t] = (x[t-1] + delta) then ID(x[t])=ID(x[t-1]) 
 otherwise ID(x[t]) = ID(x[t-1]) + 1.
}
\examples{
 \dontrun{

 x <- c(0,1,1,1,2,2,1,3,1,0,0)
 y <- c(1998:2000,2005:2010)

 make_series_id(y, delta=1)
 make_series_id(x)

 } 



}
