% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMDVEVID.R
\name{makeMDVEVID}
\alias{makeMDVEVID}
\title{makeMDVEVID()}
\usage{
makeMDVEVID(ds, DV, AMT)
}
\arguments{
\item{ds}{NONMEM formatted data.frame to add MDV/EVID column to}

\item{DV}{column number containing DV record}

\item{AMT}{column number containing AMT record}
}
\description{
This function adds MDV and EVID columns to NONMEM
formatted data.frames
}
\examples{
makeMDVEVID()
}
\keyword{dataTools}
