% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeByDate.R
\name{mergeByDate}
\alias{mergeByDate}
\title{Merge by Date}
\usage{
mergeByDate(ds, ds2add)
}
\arguments{
\item{ds}{parent data.frame to be added to and return}

\item{ds2add}{data to pull from}
}
\description{
This function converts patient data in wide form to long form
}
\examples{
mergeByDate()
}
\keyword{dataTools}
