% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeByDate.R
\name{mergeByDate}
\alias{mergeByDate}
\title{Merge by Date}
\usage{
mergeByDate(ds, ds2add)
}
\arguments{
\item{ds}{Data.frame to be added to and returned.}

\item{ds2add}{Data to pull covariates from and merge into \code{ds}. The
function assumes this data.frame has 2 columns named \code{SID} and
\code{DATE} for merging, with a third column with the variable to be
included.}
}
\value{
Data.frame \code{ds} with an added column from \code{ds2add}.
}
\description{
mergeByDate function adds new column to ds that contains VAR from nearest
past DATE for SID to each row and returns ds.
}
\details{
Long description.
}
\examples{



}
\author{
Samuel Callisto \email{calli055@umn.edu} mergeByDate()
}
