% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericTime.R
\name{numericTime}
\alias{numericTime}
\title{numericTime()}
\usage{
numericTime(vec)
}
\arguments{
\item{vec}{vector containing time values of format HH:MM stored as strings}
}
\description{
This function converts clock times stored as strings
into numeric hours
}
\examples{
numericTime()
}
\keyword{dataTools}
