% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered_factor.R
\name{ordered_factor}
\alias{ordered_factor}
\title{Creates an ordered factor based on order of another variable}
\usage{
ordered_factor(var, seq)
}
\arguments{
\item{var}{variable to be converted into a factor}

\item{seq}{variable that implies the sequence of \code{var} when the function \code{order} is applied to it}
}
\value{
\code{factor(var)} with the implied order by \code{seq}.
}
\description{
Creates an ordered factor based on order of another variable
}
\details{
This function helps to re-order groups in a \code{ggplot2} plot. See example below.
}
\examples{

val <- rpois(10,100)
num <- sample(seq(1,10))
	
levels(ordered_factor(val, num))

}
