% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Relationship.R, R/sugar.R
\name{Relationship}
\alias{Relationship}
\alias{Rel}
\title{Relationship Class}
\usage{
Rel(left = NULL, right = NULL)
}
\arguments{
\item{left}{Optional object of class \code{data.table} (or \code{data.frame}, see Details) that is treated as the left table.}

\item{right}{Optional object of class \code{data.table} (or \code{data.frame}, see Details) that is treated as the right table.}
}
\value{
A \code{Relationship} object.
}
\description{
A class for relationships between objects of class \code{data.frame}.
}
\details{
If the objects passed are not \code{data.table} objects then a \code{data.table} copy of the objects will be created with \code{data.table::as.data.table}.
}
\section{Functions}{
\itemize{
\item \code{Rel()}: Relationship$new() constructor sugar

}}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{left}}{The left table.}

\item{\code{right}}{The right table.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Relationship-new}{\code{Relationship$new()}}
\item \href{#method-Relationship-get_on}{\code{Relationship$get_on()}}
\item \href{#method-Relationship-on}{\code{Relationship$on()}}
\item \href{#method-Relationship-is_not_fully_specified}{\code{Relationship$is_not_fully_specified()}}
\item \href{#method-Relationship-clone}{\code{Relationship$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-new"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-new}{}}}
\subsection{Method \code{new()}}{
Relationship constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$new(left = NULL, right = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{left}}{Optional object of class \code{data.table} (or \code{data.frame}, see Details) that is treated as the left table.}

\item{\code{right}}{Optional object of class \code{data.table} (or \code{data.frame}, see Details) that is treated as the right table.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-get_on"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-get_on}{}}}
\subsection{Method \code{get_on()}}{
Get the relationship connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$get_on()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The connection.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-on"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-on}{}}}
\subsection{Method \code{on()}}{
Specify the columns the relationship between left and right is based on.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$on(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Which columns in the left table relate to which columns of the right table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns itself.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-is_not_fully_specified"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-is_not_fully_specified}{}}}
\subsection{Method \code{is_not_fully_specified()}}{
Check if the relationship is not fully specified.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$is_not_fully_specified()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if \code{left}, \code{right} and \code{on} are specified and \code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
