% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Relationship.R
\name{Relationship}
\alias{Relationship}
\title{Relationship Class}
\value{
A \code{Relationship} object.
}
\description{
A class for relationships between objects of class \code{data.frame}.
}
\details{
If the objects passed are not \code{data.table} objects then a \code{data.table} copy of the objects will be created with \code{data.table::as.data.table}.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{left}}{The left table.}

\item{\code{right}}{The right table.}

\item{\code{on}}{The on clause.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Relationship-new}{\code{Relationship$new()}}
\item \href{#method-Relationship-print}{\code{Relationship$print()}}
\item \href{#method-Relationship-reverse}{\code{Relationship$reverse()}}
\item \href{#method-Relationship-is_special}{\code{Relationship$is_special()}}
\item \href{#method-Relationship-right_on}{\code{Relationship$right_on()}}
\item \href{#method-Relationship-is_not_fully_specified}{\code{Relationship$is_not_fully_specified()}}
\item \href{#method-Relationship-clone}{\code{Relationship$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-new"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-new}{}}}
\subsection{Method \code{new()}}{
Relationship constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$new(left = NULL, right = NULL, on)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{left}}{Optional object of class \code{data.table} (or \code{data.frame}, see Details) that is treated as the left table.}

\item{\code{right}}{Optional object of class \code{data.table} (or \code{data.frame}, see Details) that is treated as the right table.}

\item{\code{on}}{The \code{on} clause specifying matching column names in left and right.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-print"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-reverse"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-reverse}{}}}
\subsection{Method \code{reverse()}}{
Reverse the left and right tables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$reverse()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new Relationship object with reversed tables.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-is_special"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-is_special}{}}}
\subsection{Method \code{is_special()}}{
Check if relationship is special.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$is_special()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical TRUE or FALSE.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-right_on"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-right_on}{}}}
\subsection{Method \code{right_on()}}{
Relating columns from the right table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$right_on()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector of column names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-is_not_fully_specified"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-is_not_fully_specified}{}}}
\subsection{Method \code{is_not_fully_specified()}}{
Check if the relationship is not fully specified.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$is_not_fully_specified()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if \code{left}, \code{right} and \code{on} are specified and \code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Relationship-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Relationship-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Relationship$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
