% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize2.R
\name{standardize2}
\alias{standardize2}
\title{Standardize a Vector}
\usage{
standardize2(x, scale = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{scale}{scale variable by twice its standard deviation?}
}
\value{
standardized \code{vector}.
}
\description{
Standardize a Vector
}
\details{
The function mean-centers values in the vector and scale them by twice their standard deviation. 
The function ignores all missing values.
}
\examples{
 \dontrun{

 x <- runif(100, 0, 1.2)
 xz <- standardize2(x)
 sd(xz)
 
 }

}
\references{
Gelman, A. (2008). Scaling Regression Inputs by Dividing by two Standard Deviations. Statistics in Medicine 27(2865–2873).
}
