% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatFrame.R
\name{StatFrame}
\alias{StatFrame}
\title{Data aggregation}
\description{
Data aggregation

Data aggregation
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StatFrame-new}{\code{StatFrame$new()}}
\item \href{#method-StatFrame-count}{\code{StatFrame$count()}}
\item \href{#method-StatFrame-aggregate}{\code{StatFrame$aggregate()}}
\item \href{#method-StatFrame-sum}{\code{StatFrame$sum()}}
\item \href{#method-StatFrame-mean}{\code{StatFrame$mean()}}
\item \href{#method-StatFrame-median}{\code{StatFrame$median()}}
\item \href{#method-StatFrame-min}{\code{StatFrame$min()}}
\item \href{#method-StatFrame-max}{\code{StatFrame$max()}}
\item \href{#method-StatFrame-quantile}{\code{StatFrame$quantile()}}
\item \href{#method-StatFrame-apply}{\code{StatFrame$apply()}}
\item \href{#method-StatFrame-summary}{\code{StatFrame$summary()}}
\item \href{#method-StatFrame-copy}{\code{StatFrame$copy()}}
\item \href{#method-StatFrame-clone}{\code{StatFrame$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatFrame-new"></a>}}
\if{latex}{\out{\hypertarget{method-StatFrame-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatFrame$new(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatFrame-count"></a>}}
\if{latex}{\out{\hypertarget{method-StatFrame-count}{}}}
\subsection{Method \code{count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatFrame$count(by = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatFrame-aggregate"></a>}}
\if{latex}{\out{\hypertarget{method-StatFrame-aggregate}{}}}
\subsection{Method \code{aggregate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatFrame$aggregate(funs, columns = NULL, where = NULL, by = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{funs}}{A list of functions used to create an aggregate summary.}

\item{\code{columns}}{Optional list of columns to aggregate or a predicate.}

\item{\code{where}}{Filtering}

\item{\code{by}}{By group.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatFrame-sum"></a>}}
\if{latex}{\out{\hypertarget{method-StatFrame-sum}{}}}
\subsection{Method \code{sum()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatFrame$sum(columns = is.numeric, where = NULL, by = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatFrame-mean"></a>}}
\if{latex}{\out{\hypertarget{method-StatFrame-mean}{}}}
\subsection{Method \code{mean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatFrame$mean(columns = is.numeric, where = NULL, by = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatFrame-median"></a>}}
\if{latex}{\out{\hypertarget{method-StatFrame-median}{}}}
\subsection{Method \code{median()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatFrame$median(columns = is.numeric, where = NULL, by = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatFrame-min"></a>}}
\if{latex}{\out{\hypertarget{method-StatFrame-min}{}}}
\subsection{Method \code{min()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatFrame$min(columns = is.numeric, where = NULL, by = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatFrame-max"></a>}}
\if{latex}{\out{\hypertarget{method-StatFrame-max}{}}}
\subsection{Method \code{max()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatFrame$max(columns = is.numeric, where = NULL, by = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatFrame-quantile"></a>}}
\if{latex}{\out{\hypertarget{method-StatFrame-quantile}{}}}
\subsection{Method \code{quantile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatFrame$quantile(probs, columns = is.numeric, where = NULL, by = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatFrame-apply"></a>}}
\if{latex}{\out{\hypertarget{method-StatFrame-apply}{}}}
\subsection{Method \code{apply()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatFrame$apply(fun, columns = NULL, where = NULL, by = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatFrame-summary"></a>}}
\if{latex}{\out{\hypertarget{method-StatFrame-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatFrame$summary(type = is.numeric)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatFrame-copy"></a>}}
\if{latex}{\out{\hypertarget{method-StatFrame-copy}{}}}
\subsection{Method \code{copy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatFrame$copy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatFrame-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StatFrame-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatFrame$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
